Fix issue with published setting
If archivesspace and islandora are set to different defaults for the
published state, then the object could end up published in one system
and not in the other on initial ingest. This fixes that issue by giving
the user the option to choose if the item is published or not upon
ingest.

Fixes #8