Modify REST API to include GET and POST
Before when using the REST API a PUT request to an object that wasn't
linked into archivesspace would make the object into an archivesspace
object, now it returns an error. You can link an object to an
ArchivesSpace object by calling a POST requst on it. There is also a new
GET request that lets you see if an object is linked to an ArchivesSpace
object or not.

Fixes #10.